design: 컴포넌트 위치 조정, tailwind 수정
