+++ src/components/AccessTimeRefresh/index.tsx
@@ -14,7 +14,7 @@ const AccessTimeRefresh = () => {
   }, []);
 
   return (
-    <div>
+    <div className="pr-[12px]">
       <ul className="flex items-center justify-end gap-[4px] text-gray-300 text-B-14-M">
         <li>{accessTime} 기준</li>
         <li>
